+++ app/src/main/java/com/example/househomey/ViewItemFragment.java
@@ -12,6 +12,8 @@ import android.widget.TextView;
 
 import androidx.fragment.app.Fragment;
 
+import com.example.househomey.form.EditItemFragment;
+
 /**
  * This fragment is for the "View Item Page" - which currently displays the details and comment linked
  * to the item
