+++ app/src/main/java/com/example/househomey/form/ItemFormFragment.java
@@ -1,4 +1,4 @@
-package com.example.househomey;
+package com.example.househomey.form;
 
 import static java.util.Objects.requireNonNull;
 
@@ -12,6 +12,9 @@ import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
 
+import com.example.househomey.Item;
+import com.example.househomey.MainActivity;
+import com.example.househomey.R;
 import com.google.android.material.datepicker.CalendarConstraints;
 import com.google.android.material.datepicker.DateValidatorPointBackward;
 import com.google.android.material.datepicker.MaterialDatePicker;
